Okay, final part of exercise 3 finished (I think)
All of this work really needs to be checked.  I think it's all OKAY, but I'm not sure.
Additionally, exercise1 has really awful code (I coded all of it when I had a less robust understanding of the language), and it wouldn't hurt to redo a lot of it.
Some of exercise 2 could definitely be reworked, too.